# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handler.CarResponse:
    properties:
      color:
        type: string
      createdAt:
        description: ✅ camelCase
        type: string
      id:
        type: string
      licensePlate:
        description: ✅ camelCase
        type: string
      make:
        type: string
      mileage:
        type: integer
      model:
        type: string
      ownerID:
        description: ✅ camelCase
        type: string
      updatedAt:
        description: ✅ camelCase
        type: string
      vin:
        type: string
      year:
        type: integer
    type: object
...
  handler.CreateCarRequest:
    properties:
      color:
        type: string
      licensePlate:
        type: string
      make:
        type: string
      mileage:
        type: integer
      model:
        type: string
      ownerID:
        description: 'OwnerID optional: solo personal del taller (admin/manager/employee)
          asigna el cliente dueño.'
        type: string
      vin:
        type: string
      year:
        type: integer
    type: object
...
  handler.SwaggerMessage:
    properties:
      details:
        type: string
      error:
        type: string
      message:
        type: string
    type: object
...
  handler.UpdateCarRequest:
    properties:
      color:
        type: string
      licensePlate:
        description: ✅ camelCase
        type: string
      make:
        type: string
      mileage:
        type: integer
      model:
        type: string
      vin:
        type: string
      year:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 'API de gestión de taller: autenticación JWT, coches, citas y empleados.'
  title: GonsGarage API
  version: "1.0"
paths:
...
  /api/v1/cars:
    get:
      parameters:
      - description: UUID del cliente dueño (solo staff)
        in: query
        name: ownerId
        type: string
      - description: Límite (staff sin ownerId; default 50)
        in: query
        name: limit
        type: integer
      - description: Offset paginación
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.CarResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Listar coches
      tags:
      - cars
    post:
      consumes:
      - application/json
      parameters:
      - description: Datos del vehículo
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateCarRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.CarResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Crear coche
      tags:
      - cars
  /api/v1/cars/{id}:
    delete:
      parameters:
      - description: UUID del coche
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Sin cuerpo
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Eliminar coche
      tags:
      - cars
    get:
      parameters:
      - description: UUID del coche
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CarResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Obtener coche
      tags:
      - cars
    put:
      consumes:
      - application/json
      parameters:
      - description: UUID del coche
        in: path
        name: id
        required: true
        type: string
      - description: Campos a actualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateCarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CarResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.SwaggerMessage'
      security:
      - BearerAuth: []
      summary: Actualizar coche
      tags:
      - cars
...
securityDefinitions:
  BearerAuth:
    description: 'JWT: cabecera Authorization con valor Bearer seguido del token (rutas
      bajo /api/v1 salvo /health).'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"